Changed how input behaves and is managed
Due to how std::unique_ptr works, when a new input was created, the
object was immediately destroyed by the destructor of the old one.
Now input is held by the application as a normal pointer, and to avoid
leaks, it is checked on the contructor of the Input for other instances.